+++ wish-models/src/wish_models/__init__.py
@@ -4,6 +4,7 @@ from .command_result.log_files import LogFiles
 from .executable_collection import ExecutableCollection
 from .executable_info import ExecutableInfo
 from .knowledge.knowledge_metadata import KnowledgeMetadata, KnowledgeMetadataContainer
+from .settings import Settings, settings
 from .system_info import SystemInfo
 from .utc_datetime import UtcDatetime
 from .wish.wish import Wish
@@ -22,4 +23,6 @@ __all__ = [
     "SystemInfo",
     "ExecutableInfo",
     "ExecutableCollection",
+    "Settings",
+    "settings",
 ]