+++ CHANGELOG.md
@@ -9,6 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Fixed
+
+### Removed
+
+
+## [0.3.1] - 2025-03-09
+
+### Added
+
 - Added support for custom .env file path:
   - Added `--env-file` command-line option to specify a custom .env file path
   - Default path is `$WISH_HOME/env`